+++ python/metatomic_ase/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to metatomic-ase are documented here, following the
 [keep a changelog](https://keepachangelog.com/en/1.1.0/) format. This project
 follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- Possible sections for each package:
+<!-- Possible sections
 ### Added
 
 ### Fixed